refactor: Utilizando expansão do mocktail
